Implementing Generalized linear mixed models
I have added the variable "model" to the function linmixmod, which allows to perfom either a linear mixed model (model=0) or a generalized linear mixed model (GLMM) with a log-link function (model=1). GLMM might be better suited for skewed distributions like deformation. The user should be able to decide which model should be applied. Therefore I suggest a dropdown menu for LMM and GLMM