Perform log-scaling before computing density
Currently, the density plots look very awkward when plotting in log-space. This is because the log-transformation changes the spacing between positions. The proposed change introduces a "scaling" function, which is used to perform the scaling for x and y or both if necessary. The kde-accuracies are adjusted as well (xacc and yacc)